third_party/WebKit/LayoutTests/fast/hidpi/image-srcset-nomodifier-expected.txt

PASS document.getElementById("foo").clientWidth==150 is true
PASS successfullyParsed is true

TEST COMPLETE
This test passes if the image below is not empty. It ensures that a candidate without scale modifier get a default one
